fix: 계산 결과 타입 double로 수정

+++ src/main/java/com/programmers/Calculator.java
@@ -4,8 +4,8 @@ import java.util.ArrayDeque;
 import java.util.Deque;
 
 public class Calculator {
-    public int calculate(String postfixExpression) {
-        Deque<Integer> operandStack = new ArrayDeque<>();
+    public double calculate(String postfixExpression) {
+        Deque<Double> operandStack = new ArrayDeque<>();
 
         postfixExpression.chars()
                 .mapToObj(token -> (char) token)
@@ -20,17 +20,17 @@ public class Calculator {
         return operandStack.pop();
     }
 
-    private static void handleDigit(Deque<Integer> operandStack, char digit) {
-        int operand = Character.getNumericValue(digit);
+    private static void handleDigit(Deque<Double> operandStack, char digit) {
+        double operand = Character.getNumericValue(digit);
         operandStack.push(operand);
     }
 
-    private static void handleOperator(Deque<Integer> operandStack, char symbol) {
-        int operand2 = operandStack.pop();
-        int operand1 = operandStack.pop();
+    private static void handleOperator(Deque<Double> operandStack, char symbol) {
+        double operand2 = operandStack.pop();
+        double operand1 = operandStack.pop();
 
         Operator operator = Operator.findBySymbol(symbol);
-        int result = operator.calculate(operand1, operand2);
+        double result = operator.calculate(operand1, operand2);
 
         operandStack.push(result);
     }
